Traverse local drive and filter files to be downloaded
I use pathlib, os to traverse local drive, find out which files need to be downloaded. I also compare the timestamps of remote and local versions, and download it if the remote version has a later timestamp. However, Qiniu seems to have made a mistake on their API page saying the unit of their timestamp is 100ns. The correct unit seems to be 1us.