Add batch upload function, modify filter logic
The program can now upload file to remote Qiniu drive with the same Qiniu API. It relies on python mimetypes library to guess the mime_type of the file (which in turn uses file extension to determine mime_type). The filter logic is modified so that it doesn't upload a local file when its timestamp is later than the remote counterpart.