Traverse local files, compare with remote and upload files
I used os.walk function to traverse a local file tree and find out the files that exist locally but not remotely.

One problem remains, when I upload a file from local to remote, its remote upload time is going to *later* than the original modify timestamp. When the program is run again, it is going to look like the local one is an older version... I will try to use some trick to reset its ST_MTIME to now.