Add feature to synch multiple Buckets
Class "MultiBackupDriver" is created to conduct multiple backup. There is some problem with the argument "options" because QiniuBackup expects a configparser.SectionProxy, with the useful function getboolean() which translates phrase like "True" to True. I have to write a mock derived class of dict to keep it working. I don't like this very much but at least it works.