Change config file to TOML
TOML is superior because it is  able to parse into multiple data types, not just str. I utilized the Array of Table feature and reorganized the config file as I always wanted to. I also delayed the look-up of keys into subclasses instead of taking it all at the MultiQbackupDriver level.